[Upgrade NanoHTTPD] applying to WebSocket part
