+++ nodel-jyhost/src/main/java/org/nodel/websockets/QueueEntry.java
@@ -1,0 +1,18 @@
+package org.nodel.websockets;
+
+public class QueueEntry {
+    /**
+     * The websocket data.
+     */
+    public WebSocketMessage msg;
+
+    /**
+     * 'Ping' to be sent
+     */
+    public boolean ping;
+
+    /**
+     * Socket to close because of silence.
+     */
+    public boolean silentTooLong;
+}
